+++ src/main/java/vn/kms/ngaythobet/web/dto/CreateMatchInfo.java
@@ -5,26 +5,31 @@ package vn.kms.ngaythobet.web.dto;
 import java.time.LocalDateTime;
 
 import org.hibernate.validator.constraints.NotEmpty;
+import org.springframework.format.annotation.DateTimeFormat;
 
 import vn.kms.ngaythobet.domain.tournament.Competitor;
 import vn.kms.ngaythobet.domain.validation.EntityExist;
+import vn.kms.ngaythobet.domain.validation.FieldNotMatch;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
+@FieldNotMatch(firstField="competitor1", secondField="competitor2", message="{validation.matches.fieldNotMatch.message}")
 public class CreateMatchInfo {
     @NotNull
+    @EntityExist(type = Competitor.class, message="{validation.matches.existRoundEntity.message}")
     private Long round;
 
     @NotNull
-    @EntityExist(type = Competitor.class)
+    @EntityExist(type = Competitor.class, message="{validation.matches.existCompetitorEntity.message}")
     private Long competitor1;
     
     @NotNull
-    @EntityExist(type = Competitor.class)
+    @EntityExist(type = Competitor.class, message="{validation.matches.existCompetitorEntity.message}")
     private Long competitor2;
     
     @NotNull
+    @DateTimeFormat(pattern="yyyy-MM-dd hh:mm a")
     private LocalDateTime time;
     
     @NotEmpty